requirements.txt
Django==1.4
South==0.7.5
argparse==1.2.1
chardet==2.1.1
dj-database-url==0.2.1
django-appconf==0.6
django-compressor==1.1.2
django-csvimport==0.9
django-debug-toolbar==0.9.4
django-devserver==0.3.2
psycopg2==2.4.5
six==1.2.0
wsgiref==0.1.2
django-tastypie==0.9.14
simplejson==3.1.3
django-ckeditor==5.0.3
django-aloha-editor==1.0
